Refactor Lobby into smaller components.
