+++ next/src/styles/NameForm.module.css
@@ -1,0 +1,28 @@
+.nameForm {
+    padding: 0;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-around;
+    width: 100%;
+}
+
+.nameForm * {
+    font-size: 1.5rem;
+}
+
+.nameForm p {
+    color: red;
+    margin: 0;
+    margin-top: 1rem;
+}
+
+@media (min-width: 1200px) {
+    .nameForm {
+        align-items: center;
+    }
+
+    .nameForm label {
+        text-align: left;
+        min-width: 100%;
+    }
+}
