--- next/src/styles/NameModal.module.css
@@ -1,14 +1,0 @@
-.nameModalBackground {
-    height: 100%;
-    width: 100%;
-    position: fixed;
-    background-color: black;
-}
-
-.nameModal {
-    height: 50%;
-    width: 50%;
-    background-color: blue;
-    color: white;
-    border-color: white;
-}
